feat : spacings + extract SingleHeader & Single Test as internal components + check form validity
