+++ src/views/Run/internal/Configuration/internal/HeadersConfiguration/internal/HeadersConfiguration.types.ts
@@ -1,6 +1,1 @@
-export type Header = {
-  key: string
-  value: string
-}
-
 export type Headers = Record<string, string[]>
