+++ src/views/Run/internal/Configuration/internal/HeadersConfiguration/internal/SingleHeader/internal/SingleHeader.types.ts
@@ -1,0 +1,4 @@
+export type Header = {
+  key: string
+  value: string
+}
