+++ src/views/Run/internal/Configuration/internal/HeadersConfiguration/internal/SingleHeader/internal/single-header.module.scss
@@ -1,0 +1,5 @@
+.single-header {
+  &__value {
+    min-width: none;
+  }
+}
