+++ src/views/Run/internal/Configuration/internal/TestsConfiguration/internal/SingleTest/index.ts
@@ -1,0 +1,1 @@
+export { SingleTest } from './SingleTest'
