+++ src/views/Run/internal/Configuration/internal/TestsConfiguration/internal/TestsConfiguration.helpers.ts
@@ -1,110 +1,13 @@
-import { ChangeEventHandler, MouseEventHandler } from 'react'
+import { MouseEventHandler } from 'react'
 
-import { GoDuration } from '@/benchttp/common'
 import { ConfigurationTestCase } from '@/benchttp/configuration'
-import {
-  isDurationMetricField,
-  isMetricField,
-  isNumberMetricField,
-} from '@/benchttp/metrics'
-import { parseInteger } from '@/tools/parsers'
-
-import s from './TestsConfiguration.module.scss'
-
-export const handleNameChange = (
-  tests: ConfigurationTestCase[],
-  index: number,
-  onChange: (tests: ConfigurationTestCase[]) => void
-): ChangeEventHandler<HTMLInputElement> => {
-  return (e) => {
-    const newTests = [...tests]
-    newTests[index].name = e.target.value
-    onChange(newTests)
-  }
-}
-
-export const handleFieldChange = (
-  tests: ConfigurationTestCase[],
-  index: number,
-  onChange: (tests: ConfigurationTestCase[]) => void
-): ChangeEventHandler<HTMLInputElement> => {
-  return (e) => {
-    const newTests = [...tests]
-    // @ts-expect-error e.target.value is expected to be any string,
-    // a invalid value will set the input state to invalid.
-    newTests[index].field = e.target.value
-    onChange(newTests)
-  }
-}
-
-export const isValidTestField = (field: string) => isMetricField(field)
-
-export const handlePredicateChange = (
-  tests: ConfigurationTestCase[],
-  index: number,
-  onChange: (tests: ConfigurationTestCase[]) => void
-): ChangeEventHandler<HTMLSelectElement> => {
-  return (e) => {
-    const newTests = [...tests]
-    // TODO: fix this type assertion, it's not safe at runtime.
-    // @ts-expect-error e.target.value is a TestPredicate because
-    // the select options are limited to TestPredicate values.
-    newTests[index].predicate = e.target.value
-    onChange(newTests)
-  }
-}
-
-export const isValidTestTarget = (
-  target: number | GoDuration,
-  field: string
-): boolean => {
-  switch (true) {
-    case isDurationMetricField(field):
-      try {
-        return (
-          typeof target === 'string' &&
-          target.endsWith('ms') &&
-          Number.isInteger(parseInteger(target))
-        )
-      } catch {
-        return false
-      }
-    case isNumberMetricField(field):
-      return Number.isInteger(target)
-    default:
-      return false
-  }
-}
-
-export const handleDurationTargetChange = (
-  tests: ConfigurationTestCase[],
-  index: number,
-  onChange: (tests: ConfigurationTestCase[]) => void
-): ((value: `${number}ms`) => void) => {
-  return (value) => {
-    const newTests = [...tests]
-    newTests[index].target = value
-    onChange(newTests)
-  }
-}
-
-export const handleNumberTargetChange = (
-  tests: ConfigurationTestCase[],
-  index: number,
-  onChange: (tests: ConfigurationTestCase[]) => void
-): ((value: number) => void) => {
-  return (value) => {
-    const newTests = [...tests]
-    newTests[index].target = value
-    onChange(newTests)
-  }
-}
+import { isNumberMetricField } from '@/benchttp/metrics'
 
 const placeholderTest = (): ConfigurationTestCase => ({
   name: '',
   field: 'ResponseTimes.Max',
   predicate: 'LT',
-  target: '200ms',
+  target: '0ms',
 })
 
 const t = placeholderTest()
@@ -136,24 +39,4 @@ export const handleRemoveTest = (
     if (!enabled) return
     onChange(tests.filter((_, i) => i !== index))
   }
-}
-
-export const getIconClassNames = ({
-  className,
-  enabled,
-}: {
-  className?: string
-  enabled: boolean
-}): string[] => {
-  const classNames: string[] = []
-
-  if (className) {
-    classNames.push(className)
-  }
-
-  if (!enabled) {
-    classNames.push(s['run-configuration-panel-tests__icon--disabled'])
-  }
-
-  return classNames
 }
